Moved middleware function out of multiplayer client, moved MP code to its own folder as with audio
