+++ services/app/src/components/multiplayer/MultiplayerFooterContainer.tsx
@@ -2,12 +2,12 @@ import {connect} from 'react-redux';
 import Redux from 'redux';
 import {setDialog} from '../../actions/Dialog';
 import {syncMultiplayer} from '../../actions/Multiplayer';
-import {getMultiplayerClient} from '../../Multiplayer';
+import {getMultiplayerConnection} from '../../multiplayer/Connection';
 import {AppState, DialogIDType} from '../../reducers/StateTypes';
 import MultiplayerFooter, {DispatchProps, Props, StateProps} from './MultiplayerFooter';
 
 const mapStateToProps = (state: AppState, ownProps: Partial<Props>): StateProps => {
-  const rpClient = getMultiplayerClient();
+  const rpClient = getMultiplayerConnection();
   return {
     multiplayer: state.multiplayer,
     cardTheme: ownProps.cardTheme || 'light',
