+++ services/app/src/components/views/quest/cardtemplates/combat/NoTimerContainer.tsx
@@ -1,4 +1,4 @@
-import {getMultiplayerClient} from 'app/Multiplayer';
+import {getMultiplayerConnection} from 'app/multiplayer/Connection';
 import {SettingsType} from 'app/reducers/StateTypes';
 import {connect} from 'react-redux';
 import Redux from 'redux';
@@ -13,7 +13,7 @@ import {mapStateToProps} from './Types';
 const mapDispatchToProps = (dispatch: Redux.Dispatch<any>): DispatchProps => {
   return {
     onTimerStop: (node: ParserNode, settings: SettingsType, elapsedMillis: number, surge: boolean, seed: string) => {
-      const multiplayerConnected = getMultiplayerClient().isConnected();
+      const multiplayerConnected = getMultiplayerConnection().isConnected();
 
       // We don't want to **stop** the timer if we're connected to remote
       // play. Rather, we want to wait until everyone's timer is stopped
